rctype_Road IterationStepSize formula fixes
IteratedGeometry3D_0.Separation is deprecated and now named IteratedGeometry3D_0.IterationStepSize .

There was also a mismatch between the constructor name '_JBTKU_VEI_IteratedGeometrySeparation' and the function name '_JBTKU_VEI_IteratedGeometry3DSeparation', which caused an infinite loop in the property formula. I've updated all names to match the new property name and to be the same.